Fix JDK11 test failure by ensuring class not a nestmember
PrivateMethod was declared as an inner class of AdaptorTests
which makes it a nest member in JDK11.  Nest members have private
access in JDK11+ which defeats the point of the nest.

This change makes the two classes siblings rather than nestmates.
